feat(contract-integration): re-engineered initial loading flow
