+++ src/index.tsx
@@ -4,21 +4,11 @@ import App from "./App";
 import "./i18n";
 import "bootstrap/dist/css/bootstrap.css";
 import "react-toastify/dist/ReactToastify.css";
-import {Provider} from "react-redux";
-import {configureStore} from "./store";
-import startDataPuller from "./common/utils/data-puller";
-import subscribeOnEvents from "./common/utils/subscriber";
 // import * as serviceWorker from './serviceWorker';
-
-const store = configureStore();
-startDataPuller(store);
-subscribeOnEvents(store);
 
 ReactDOM.render(
     <React.StrictMode>
-        <Provider store={store}>
-            <App />
-        </Provider>
+        <App />
     </React.StrictMode>,
     document.getElementById("root")
 );
